research-app, research-app-messages: Move SelectionStateMessage and the exported source interface into selections.ts, and make the appropriate changes in index.ts and in the research app.
